+++ src/core/e3_interface.cpp
@@ -189,7 +189,7 @@ void E3Interface::setup_loop() {
         }
         
         // Decode the setup request
-        auto decode_result = encoder_->decode(buffer.data(), ret);
+        auto decode_result = encoder_->decode(buffer.data(), static_cast<size_t>(ret));
         if (!decode_result) {
             E3_LOG_ERROR(LOG_TAG) << "Failed to decode setup request";
             continue;
